Refactor models
