+++ flare_portal/experiments/models/__init__.py
@@ -1,0 +1,12 @@
+from .core import Experiment, Participant, Project
+from .data import FearConditioningData
+from .modules import BaseModule, FearConditioningModule
+
+__all__ = [
+    "BaseModule",
+    "Experiment",
+    "FearConditioningData",
+    "FearConditioningModule",
+    "Participant",
+    "Project",
+]
